Interpolate <parent> element in the interpolated/resolved model
If you are using variables in the parent element version, you need to
interpolate it in the flattened POM. Of course, that's only when you
need to keep the connection to the parent. One reason to do so is to
keep profiles inherited.

In this case you need to keep the parent element and set the
ElementHandling to "interpolate" explicitly:

<pomElements>
  <parent>interpolate</parent>
</pomElements>